Update jsoneditor.barebones-theme.js
Barebones theme previously relied on the theme superclass to create its elements.  This changes that so the theme itself creates the elements.  This way the super's extra inline styling is not used.  Additionally, certain tags and attributes have been changes to be more WAI/HTML5 appropriate.

Pending change:  I am not aware which classes from the original framework should be preserved.