Improve performance of theme.setButtonText
Profiling showed that for large object trees, a significant portion of time was being spent in setButtonText.
Removing the `innerHTML += ' '` fixes this.

I also use a better-performing strategy for removing the contents of the button, although that's probably a premature optimization.

Unfortunately, in the grand scheme of things, this still makes a barely-noticeable difference in the time it takes to set values in the editor.