Update validator.js
As user I want to see human-readable error message if a requierd property is not populated. 
Act: error text with the property name as the reference to the failure field
Exp: error test with the property title if exists or otherwise property  name, as the reference to the failure field

This problem is very actual for multi-language projects because user might be not able to match English name of the property and form field label on his own language.