Change from static data to dynamic discovery
This change removes the data-index item, and instead discovers the index of the currently active tab - this allows
the tabs to be moved without refreshing the value - making
move, copy-in-place and drop WAY faster (not here but in
a subclass)

Also a fix to the hard-usage of the index in the editor path.
As before, instead of using the row index we use the 'editorId' set by the class/subclass. (in array.js that IS the
index, but in a subclass that, say, forgoes the use of the cache and, instead, uses editor paths NOT tied to the array
item order, the implementation of move, delete, copy, add, and drop can all be vastly improved, and in a way that does not degrade with the increasing size of the array).