Button event listeners call an instance method
All button event listeners now call an instance method for the
main implementation, allowing subclasses to implement different
strategies (e.g. see FastModArray). Those event listener MAY
choose to return a value that indicates to the event listener that
the remaining functionality of the listener should be executed.

Mostly that's specifically returning `true` - which is backward
compatible to the original array semantics. For one case, where an
object is expected, _not_ rturning that object has the same effect.